requirements.txt
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.0.1
crispy-bootstrap5==0.7
Django==4.2.4
django-crispy-forms==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mysqlclient==2.2.0
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.4
sqlparse==0.4.4
uritemplate==4.1.1
